synapse-agent/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn[standard]
requests
beautifulsoup4
playwright
google-api-python-client
pydantic
python-multipart
